feat: Added new annotation support (#92)
* feat: Initial Refactor changes

* chore: Refactored code for harness ci/cd plugin

* feat: Added new annotation support

* feat: Added new annotation for services

* feat: Updated Documentation

* feat: updated doc
