+++ plugins/harness-ci-cd/PluginConfiguation.md
@@ -11,3 +11,11 @@
 ## harness.io/cd-serviceId configuration:
 
 ![ServiceID](./src/assets/serviceId.png)
+
+## harness.io/pipelines configuration
+
+![Harness Pipeline URL](./src/assets/harness-pipeline-url.png)
+
+## harness.io/services configuration
+
+![Harness Service URL](./src/assets/harness-service-url.png)
