+++ plugins/harness-ci-cd/README.md
@@ -5,10 +5,11 @@ Website: [https://harness.io/](https://harness.io/)
 Welcome to the Harness NextGen CI/CD plugin for Backstage!
 
 ## Screenshots
-
+<img src="./src/assets/harness-new-annotation.png">
 <img src="./src/assets/ci-executions.png" />
 <img src="./src/assets/cd-executions.png" />
 <img src="./src/assets/ci-cd-executions.png" />
+
 
 ## Getting started
 
@@ -37,7 +38,7 @@ For testing purposes, you can also clone this repository to try out the plugin. 
 
 proxy:
   # ... existing proxy settings
-  '/harness':
+  '/harness/prod':
     target: 'https://app.harness.io/'
     headers:
       'x-api-key': '<YOUR PAT/SAT>'
@@ -97,8 +98,33 @@ const serviceEntityPage = (
 
 </details>
 
-4. Add required harness specific annotations to your software component's respective `catalog-info.yaml` file. Here is an example: https://github.com/harness/backstage-plugins/blob/main/examples/catalog-harness-cicd.yaml
+4. Add required harness specific annotations to your software component's respective `catalog-info.yaml` file.
 
+Here is an example: [catalog-info-new.yaml](../../examples/catalog-harness-cicd-new.yaml)
+```yaml
+apiVersion: backstage.io/v1alpha1
+kind: Component
+metadata:
+  # ...
+  annotations:
+    # optional annotation
+    harness.io/pipelines: |
+      labelA: <harness_pipeline_url>
+      labelB: <harness_pipeline_url>
+  # here labelA / labelB denotes the value you will see in dropdown in execution list. Refer screentshot 1
+
+    # optional annotation
+    harness.io/services: |
+      labelA: <harness_service_url>
+      labelB: <harness_service_url>
+spec:
+  type: service
+  # ...
+```
+
+
+#### Old Annotation
+Here is an example: https://github.com/harness/backstage-plugins/blob/main/examples/catalog-harness-cicd.yaml
 ```yaml
 apiVersion: backstage.io/v1alpha1
 kind: Component
@@ -116,6 +142,9 @@ spec:
   # ...
 ```
 
+
+
+Note: If new annotation is present then old annotation will be ignored for that particular catalog.  
 Note: Refer to [this](./PluginConfiguation.md) page on how to get these values from your Harness account.
 
 By default, the plugin will take all the pipelines inside the configured Harness project and show their executions. However, if your service has quiet a few pipelines, you can additionally configure the pipelines as well as associated services to show those specific execution details for the display.
