+++ plugins/harness-ci-cd/src/util/getHarnessToken.ts
@@ -1,0 +1,10 @@
+export function getSecureHarnessKey(key: string): string | undefined {
+  try {
+    const token = JSON.parse(decodeURI(atob(localStorage.getItem(key) || '')));
+    return token ? `Bearer ${token}` : '';
+  } catch (err) {
+    // eslint-disable-next-line no-console
+    console.log('Failed to read Harness tokens', err);
+    return undefined;
+  }
+}
